RB change: Update to how we set constraints in RBConstruction.
Previously we turned off all constraints in RBConstruction::add_scaled_matrix_and_vector()
when apply_dof_constraints was false. This change uses swap_constraints() instead. This acheives
the same thing as before if _stashed_constraints is empty (which is the default behavior), but
it is more general since it allows us to put some other possibly non-empty set of constraints
in _stashed_constraints and use those instead.